fix wrong types (#101)(patch)
### Fixed
- fixed wrong step types

+++ cg_lims/models/arnold/prep/sars_cov_2_prep/aggregate_qc.py
@@ -11,7 +11,7 @@ class AggregateQCDNACovArtifactUDF(BaseModel):
     """Aggregate QC (DNA) (Cov) v1"""
 
     sample_concentration: Optional[float] = Field(None, alias="Concentration")
-    sample_size: Optional[float] = Field(None, alias="Size (bp)")
+    sample_size: Optional[int] = Field(None, alias="Size (bp)")
 
 
 class AggregateQCDNACovFields(BaseStep):
